Fix windows and 32 bit

+++ pandas/tests/frame/methods/test_astype.py
@@ -676,7 +676,7 @@ class TestAstype:
         data = np.arange(16).reshape(4, 4)
         df = DataFrame(data)
 
-        result = df.iloc[:step1, :step2].astype("int32").astype("int64")
+        result = df.iloc[:step1, :step2].astype("int16").astype(np.intp)
         expected = df.iloc[:step1, :step2]
         tm.assert_frame_equal(result, expected)
